+++ src/main/java/com/javPOL/magazineJava/dao/CustomerDAO/CustomerDao.java
@@ -3,5 +3,5 @@ package com.javPOL.magazineJava.dao.CustomerDAO;
 import com.javPOL.magazineJava.dao.Dao;
 import com.javPOL.magazineJava.model.Customer;
 
-public interface CustomerDao extends Dao<Customer, Integer> {
+public interface CustomerDao extends Dao<Customer, Long> {
 }
